feat: Add endpoint for retrieving top tracks
This commit adds a new endpoint to the Spotify API route for retrieving the user's top tracks. The `getTopTracks` function is implemented to fetch the data from the Spotify API and return the response in JSON format. The endpoint is accessed with the `type=top-tracks` query parameter. If there are no top tracks available, an empty array is returned.