refactor: Update TopFive component to display top tracks
The TopFive component has been modified to display the user's top tracks instead of top artists. The component now fetches the data from the Spotify API using the `getTopTracks` function and renders the track information along with the album cover. The `artists` state variable has been renamed to `tracks` to reflect the change in data. This update improves the accuracy of the displayed information and provides a better user experience.